Time: 67 ms (73.10%) | Memory: 6.1 MB (11.38%) - LeetSync

+++ 897-prime-palindrome/prime-palindrome.cpp
@@ -13,7 +13,7 @@ public:
             string l = to_string(i);
             string r = l;
             reverse(begin(r),end(r));
-            long long check = stoll(l+r.substr(1));
+            int check = stoi(l+r.substr(1));
             if(check>=n and isprime(check)){
                 return check;
             }
